removed unused function and pylint
